Change FibTailRec to accept Long instead of Int

+++ src/u02/ex/Fibonacci.scala
@@ -7,9 +7,9 @@ object Fibonacci:
     case 1 | 2 => 1
     case _ => fib(n-1) + fib(n-2)
 
-  def fibTailRec(n: Int): Int =
+  def fibTailRec(n: Int): Long =
     @annotation.tailrec
-    def _fib(n: Int, acc1: Int, acc2: Int): Int = n match
+    def _fib(n: Int, acc1: Long, acc2: Long): Long = n match
       case 0 => acc1
       case 1 => acc2
       case _ => _fib(n - 1, acc2, acc1 + acc2)
